Implemented ability to spin up testbed networks
This mainly implements a Testbed class in testbed.py that provides a few
functions. In particular, it spins up the testbed topology by creating a
transit network (using a given "name" to ID its namespaces) and allows
various clients/servers to be added. The name of the Topology must be
short enough for veth pairs and namespaces to be named around them,
which for now is about 9 characters as an absolute maximum (I believe).

The main.py program is configured as a robust CLI interface. Currently
the following commands are implemented:
 - topology <name> up
 - topology <name> down

Future "topology" commands--and "run" commands--will be implemented
soon.