Added build script for mahimahi-gen container
This container is used to capture websites into protobuf .save files a
la mahimahi; it uses a mitmproxy container that has access to the
scripts in the run/ directory. These scripts utilize the python file
`mahi_pb2.py`, which was generated by the protobuf compiler, and exposes
the RequestResponse structure. The `mitmproxy2mahimahi.py` script is
used my mitmproxy to populate this structure and save it into the .save
files. The `parse_mahimahi.py` script is also called that simply creates
a `pp_sorted.txt` file that shows the HTTP request/response headers in
plaintext. This script also can be used in a standalone manor to view
.save files.

To build the container, `build.sh` is used. A result is the Docmer image
`mahimahi-gen`. python scripts that utilize this container will soon
follow