Added mitmproxy.py for site capture
This can be utilized in the CLI with the "site" subcommand (e.x.:
`main.py site www.wikipedia.org capture`.

This utilizes the mahimahi-gen container and the custom browsertime
container to capture a website. It creates a directory in sites/ with a
protobuf_files/ directory (filled with the .save files and a readable
pp_sorted.txt file)

mitmproxy fix?